Worked on Inventory UI and Insertion
Reworked the UI of the inventory system. Added in informative borders, relocated attribute value UI elements, and changed color scheme of Inventory from Light to Dark.

Created the insertion function, elements are inserted with the same species of fish, with the order being by rarity and ID value. Time complexity is O(n) but is essentially instant with such small data sets.